requirements.txt
selenium==4.34.2
beautifulsoup4==4.13.4
pandas==2.3.1
webdriver-manager==4.0.2
requests==2.32.4
openpyxl==3.1.5
pyyaml==6.0.2